toResponse 원상복구
