+++ src/main/kotlin/com/b2/backoffice/domain/comment/service/CommentServiceImpl.kt
@@ -3,7 +3,6 @@ package com.b2.backoffice.domain.comment.service
 import com.b2.backoffice.domain.comment.dto.CommentResponse
 import com.b2.backoffice.domain.comment.dto.CommentCreateRequest
 import com.b2.backoffice.domain.comment.dto.CommentUpdateRequest
-import com.b2.backoffice.domain.comment.dto.toResponse
 import com.b2.backoffice.domain.comment.model.CommentEntity
 import com.b2.backoffice.domain.comment.repository.CommentRepository
 import com.b2.backoffice.domain.exception.ModelNotFoundException
@@ -93,3 +92,12 @@ class CommentServiceImpl(
     }
 }
 
+
+fun CommentEntity.toResponse(): CommentResponse {
+    return CommentResponse(
+        id = id,
+        content = content,
+        nickname = user.nickName,
+        createdAt = createAt
+    )
+}
